fix(TextField): separated the logic and jsx the component
